# util: faster GetExtFromCert using ExtensionsMap.