--- a/go.mod
+++ b/go.mod
@@ -3,7 +3,7 @@ module github.com/zmap/zlint
 require (
 	github.com/sirupsen/logrus v1.3.0
 	github.com/weppos/publicsuffix-go v0.4.0
-	github.com/zmap/zcrypto v0.0.0-20190726170612-6da6ff3e2049
+	github.com/zmap/zcrypto v0.0.0-20190729165852-9051775e6a2e
 	golang.org/x/crypto v0.0.0-20190701094942-4def268fd1a4
 	golang.org/x/net v0.0.0-20190404232315-eb5bcb51f2a3
 	golang.org/x/text v0.3.0