… go.sum
… vendor/github.com/zmap/zcrypto/x509/x509.go
… vendor/modules.txt