The [`github.com/zmap/zcrypto`](https://github.com/zmap/zcrypto/) library that provides certificate parsing for zlint has been [updated to add an `ExtensionsMap` field](https://github.com/zmap/zcrypto/commit/9051775e6a2e3a89ec27977077b09f4496febecf) on parsed Certificates.

In this PR the zlint `util.GetExtFromCert` function is updated to use the `ExtensionsMap` field for O(1) extension access by OID instead of needing an O(n) search of the `Extensions` slice.

Running a `-cpuprofile` of the benchmarks in master gives the following `top10` in `pprof`:
```
$ cd $GOPATH/src/github.com/zmap/zlint
$ go test --run=XXX -bench=. -cpuprofile all.profile
<snipped>
$ go tool pprof all.profile
File: zlint.test
Type: cpu
Time: Jul 26, 2019 at 10:58am (EDT)
Duration: 4.13mins, Total samples = 4.79mins (115.90%)
Entering interactive mode (type "help" for commands, "o" for options)
(pprof) top10
Showing nodes accounting for 146.71s, 51.07% of 287.25s total
Dropped 522 nodes (cum <= 1.44s)
Showing top 10 nodes out of 221
      flat  flat%   sum%        cum   cum%
    45.89s 15.98% 15.98%    142.36s 49.56%  runtime.mallocgc
    31.59s 11.00% 26.97%     39.26s 13.67%  runtime.heapBitsSetType
    18.20s  6.34% 33.31%     18.20s  6.34%  runtime.nextFreeFast
    11.27s  3.92% 37.23%     20.05s  6.98%  runtime.scanobject
     9.90s  3.45% 40.68%      9.90s  3.45%  runtime.memclrNoHeapPointers
     8.31s  2.89% 43.57%      8.31s  2.89%  encoding/asn1.ObjectIdentifier.Equal
     6.92s  2.41% 45.98%     13.52s  4.71%  github.com/zmap/zlint/util.GetExtFromCert
     5.63s  1.96% 47.94%         7s  2.44%  runtime.heapBitsForAddr
     4.60s  1.60% 49.54%    130.51s 45.43%  runtime.newobject
     4.40s  1.53% 51.07%    245.67s 85.52%  github.com/zmap/zlint.BenchmarkZlint.func4
```

Notably we can see `github.com/zmap/zlint/util.GetExtFromCert` and `encoding/asn1.ObjectIdentifier.Equal` representing 4.71% and 2.89% of cumulative CPU usage respectively.

With this branch's fix applied the `top10` changes:
```
$ cd $GOPATH/src/github.com/zmap/zlint
$ go test --run=XXX -bench=. -cpuprofile all.new.profile
<snipped>
$ go tool pprof all.new.profile
File: zlint.test
Type: cpu
Time: Jul 29, 2019 at 3:38pm (EDT)
Duration: 4.21mins, Total samples = 4.89mins (116.10%)
Entering interactive mode (type "help" for commands, "o" for options)
(pprof) top10
Showing nodes accounting for 142.94s, 48.69% of 293.55s total
Dropped 513 nodes (cum <= 1.47s)
Showing top 10 nodes out of 233
      flat  flat%   sum%        cum   cum%
    48.05s 16.37% 16.37%    138.73s 47.26%  runtime.mallocgc
    28.32s  9.65% 26.02%     37.09s 12.63%  runtime.heapBitsSetType
    14.54s  4.95% 30.97%     14.54s  4.95%  runtime.nextFreeFast
    12.19s  4.15% 35.12%     21.52s  7.33%  runtime.scanobject
    10.59s  3.61% 38.73%     30.52s 10.40%  runtime.concatstrings
     9.60s  3.27% 42.00%      9.60s  3.27%  runtime.memclrNoHeapPointers
     5.60s  1.91% 43.91%      6.88s  2.34%  runtime.heapBitsForAddr
     4.85s  1.65% 45.56%    252.74s 86.10%  github.com/zmap/zlint.BenchmarkZlint.func4
     4.73s  1.61% 47.17%    118.13s 40.24%  runtime.newobject
     4.47s  1.52% 48.69%      4.47s  1.52%  runtime.memmove
```

Now there's no more `github.com/zmap/zlint/util.GetExtFromCert` or `encoding/asn1.ObjectIdentifier.Equal` nodes present in the top 10 :tada: